Update sched.h - Potential typo
In line 783, when compiling the kernel with MTK Task Turbo, task_turbo.c may not compile correctly as the variable that is used may be hidden as it requires CONFIG_MTK_SCHED_BOOST to be y. In selene_defconfig, I searched for CONFIG_MTK_SCHED_BOOST and there was no such option so I think "#ifdef CONFIG_MTK_SCHED_BOOST" might be a typo. Therefore, MTK Task Turbo might compile with the kernel successfully.